Remove DupCache
